+++ src/components/Eliza/hooks/useScrollToError.ts
